Network communication working with new system.

+++ src/config.hpp
@@ -6,7 +6,7 @@
 extern bool g_is_client;
 extern bool g_is_server;
 extern const char *g_host;
-extern int g_port;
+extern unsigned short g_port;
 
 const int g_land_width = 800;
 const int g_land_height = 600;
